lot-render/requirements.txt
bcrypt==4.1.3
cloud-sql-python-connector==1.11.0
fastapi==0.111.1
pandas==2.2.2
pg8000==1.31.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
reportlab==4.2.2
sqlalchemy==1.4.27
pymongo==4.10.0
sendgrid==6.11.0
password-strength==0.0.3.post2